Version 1.0 Commit
Remediated bug. The problem was the (i) value being passed into CheckAnswers(). This caused multiple desc.TextContents to be returned. Checking answer would render correctly and then change to incorrect due to the multiple strings being compared (expectation was one string). Fixed by declaring a variable inside the if(true) section that simply records if the answer is correct. If answer is incorrect then return the error message.